Move CartItem to discrete component; CartItem quantity now directly impacts value on change.
